+++ NEWS.md
@@ -1,3 +1,6 @@
+# mitre (development version)
+* Added PoC for ATT&CK Tactics
+
 # mitre 0.1.1
 
 * Added a `NEWS.md` file to track changes to the package.